moved logic into separate files
