+++ src/main.rs
@@ -1,42 +1,20 @@
-use clap::{Parser, Subcommand};
-use std::error::Error;
 use std::fs;
 use std::path::PathBuf;
-use walkdir::WalkDir;
-use indicatif::ProgressBar;
-use hound::{WavReader, WavWriter};
 
-const DATA_FOLDER: &str = "home/pi/data";
+use clap::{Parser, Subcommand};
+use walkdir::WalkDir;
+
+use crate::import::{Import, import_from_sd, merge_wavs};
+
+mod import;
+mod clean;
 
 #[derive(Subcommand)]
 #[clap(about = "A tool to record audio on Linux using the command line.")]
 pub enum Commands {
-    Install(Install),
     Import(Import),
 }
 
-#[derive(Parser, Debug)]
-#[clap(about =	"Install hydrophonitor on an SD card.")]
-pub struct Install {
-    /// Path to the SD card. You can find the path to the SD card by running
-    /// `lsblk` in the terminal.
-    #[clap(short, long, required = true)]
-    pub sd_card: PathBuf,
-}
-
-#[derive(Parser, Debug)]
-#[clap(about =	"Import audio from an SD card.")]
-pub struct Import {
-    /// Path to the SD card. You can find the path to the SD card by running
-    /// `lsblk` in the terminal.
-    #[clap(short, long, required = true)]
-    pub sd_card: PathBuf,
-
-    /// Path to the output folder. If not specified, the output folder will be
-    /// the current directory.
-    #[clap(short, long)]
-    pub output_folder: Option<PathBuf>,
-}
 
 #[derive(Parser)]
 #[clap(author, version, about, long_about = None)]
@@ -46,105 +24,23 @@ pub struct Cli {
     pub commands: Commands,
 }
 
-fn import_from_sd(sd_card: &mut PathBuf, output_folder: Option<PathBuf>) -> Result<(), Box<dyn Error>>{
-    let output_folder = match output_folder {
-        Some(output_folder) => output_folder,
-        None => std::env::current_dir().unwrap_or_else(|err| {
-            eprintln!("Error: {}", err);
-            std::process::exit(1);
-        })
-    };
-
-    sd_card.push(DATA_FOLDER);
-
-    let count = WalkDir::new(sd_card.clone()).into_iter().count();
-    let progress_bar = ProgressBar::new(count as u64);
-
-    for entry in WalkDir::new(sd_card.clone()) {
-        let entry = entry?; 
-        let from = entry.path();
-        let to = output_folder.join(from.strip_prefix(sd_card.clone())?);
-
-        if entry.file_type().is_dir() {
-            fs::create_dir_all(to)?;
-        } else if entry.file_type().is_file() {
-            fs::copy(from, to)?;
-        }
-        progress_bar.inc(1);
-    }
-    progress_bar.finish();
-    Ok(())
-}
-
-
-pub fn merge_wavs(input: &std::path::PathBuf, output: &std::path::PathBuf) -> Result<(), Box<dyn Error>> {
-	// Read files from input directory
-	let mut files = std::fs::read_dir(input)?
-		.filter_map(|entry| entry.ok())
-		.filter(|entry| entry.file_type().ok().map(|t| t.is_file()).unwrap_or(false))
-		.filter(|entry| entry.path().extension().unwrap_or_default() == "wav")
-		.collect::<Vec<_>>();
-
-    // If there are no wav files, return
-    if files.is_empty() {
-        println!("No wav files found in {:?}", input);
-        return Ok(());
-    }
-	// Sort files by name
-	files.sort_by(|a, b| a.file_name().cmp(&b.file_name()));
-
-	let output_name = files.first().unwrap().file_name();
-	let output_name = output_name.to_str().unwrap();
-
-	// Get wav spec from file
-	let spec = WavReader::open(files.first().unwrap().path())?.spec();
-	let mut writer = WavWriter::create(output.join(output_name), spec)?;
-
-    let progress_bar = ProgressBar::new(files.len() as u64);
-	match spec.sample_format {
-		hound::SampleFormat::Float => {
-			for file in files {
-				let mut reader = WavReader::open(file.path())?;
-				for sample in reader.samples::<f32>() {
-					writer.write_sample(sample?)?;
-				}
-                progress_bar.inc(1);
-			}
-		},
-		hound::SampleFormat::Int => {
-			for file in files {
-				let mut reader = WavReader::open(file.path())?;
-				for sample in reader.samples::<i32>() {
-					writer.write_sample(sample?)?;
-				}
-                progress_bar.inc(1);
-			}
-		},
-	}
-    progress_bar.finish();
-	writer.finalize()?;
-	Ok(())
-}
 
 fn main() {
     let commands = Cli::parse();
 
     match commands.commands {
-        Commands::Install(install) => {
-            println!("Installing hydrophonitor on SD card at {:?}", install.sd_card);
-        }
         Commands::Import(mut import) => {
-            println!("Importing audio from SD card at {:?}", import.sd_card);
+            println!("Importing audio from SD card at {:?}", import.device);
 
-            if let Some(output_folder) = &import.output_folder {
+            if let Some(output_folder) = &import.output {
                 println!("Output folder: {:?}", output_folder);
-                import_from_sd(&mut import.sd_card, Some(output_folder.clone())).unwrap_or_else(|err| {
+                import_from_sd(&mut import.device, Some(output_folder.clone())).unwrap_or_else(|err| {
                     eprintln!("Error: {}", err);
                     std::process::exit(1);
                 });
             } else {
                 println!("Output folder: current directory");
-                import_from_sd(&mut import.sd_card, None).unwrap_or_else(|err| {
+                import_from_sd(&mut import.device, None).unwrap_or_else(|err| {
                     eprintln!("Error: {}", err);
                     std::process::exit(1);
                 });
@@ -153,7 +49,7 @@ fn main() {
             // Iterate folders inside output folder. Inside each iterated folder there is
             // a folder called "audio" which contains the wav files. Merge them into a single
             // wav file and delete the "audio" folder.
-            let output_folder = match import.output_folder {
+            let output_folder = match import.output {
                 Some(output_folder) => output_folder,
                 None => std::env::current_dir().unwrap_or_else(|err| {
                     eprintln!("Error: {}", err);
